+++ example/index.ts
@@ -1,13 +1,13 @@
 import 'zone.js'
 import 'reflect-metadata'
-
-//import { enableProdMode } from "@angular/core"
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic'
+import { enableProdMode } from "@angular/core"
+
 import { AppModule } from './src/app.module'
 import { supportDocument } from 'ack-angular-fx/web-animations.min'
 
 supportDocument(document)
 
-//enableProdMode()
+enableProdMode()
 
 platformBrowserDynamic().bootstrapModule(AppModule)