+++ src/components/AckFixedElementStage.component.ts
@@ -1,0 +1,13 @@
+import { ElementRef, Component } from "@angular/core"
+import { AckApp } from "../providers/AckApp"
+
+import { string } from "./templates/ack-fixed-element-stage.pug"
+
+@Component({
+  selector:'ack-fixed-element-stage',
+  template:string
+}) export class AckFixedElementStage{
+  constructor(public AckApp:AckApp, public ElementRef:ElementRef){
+    AckApp.fixedElementStage = ElementRef
+  }
+}